feat: hack in means of saving out the file

+++ src/sqlite/worker.ts
@@ -8,6 +8,8 @@ import { isObject } from '../utils/is-object';
 
 import './sqlite3.js';
 
+declare var self: WorkerGlobalScope;
+
 (async () => {
   const sqlite3 = await sqlite3InitModule({
     print: (...args) => console.log(...args),
@@ -25,16 +27,44 @@ import './sqlite3.js';
     const batchSize = e.data.batchSize as number;
     const separateIndex = !!e.data.separateIndex;
     const useTriggers = !!e.data.useTriggers;
+    const saveFile = !!e.data.saveFile;
 
     const testFn = separateIndex
       ? useTriggers
         ? (poolUtil: OpfsSAHPoolUtil, source: string, batchSize: number) =>
-            runTestWithSeparateIndex(poolUtil, source, batchSize, true)
-        : runTestWithSeparateIndex
+            runTestWithSeparateIndex(
+              poolUtil,
+              source,
+              batchSize,
+              true,
+              saveFile
+            )
+        : (poolUtil: OpfsSAHPoolUtil, source: string, batchSize: number) =>
+            runTestWithSeparateIndex(
+              poolUtil,
+              source,
+              batchSize,
+              false,
+              saveFile
+            )
       : runTest;
 
     testFn(poolUtil, source, batchSize).then((results) => {
-      postMessage({ type: 'result', ...results });
+      let buffer: Uint8Array | null = null;
+      if (saveFile) {
+        buffer = sqlite3.capi.sqlite3_js_db_export(
+          (results as Awaited<ReturnType<typeof runTestWithSeparateIndex>>).db
+            .pointer
+        );
+      }
+
+      if (buffer) {
+        postMessage({ type: 'result', ...results, file: buffer }, [
+          buffer.buffer,
+        ]);
+      } else {
+        postMessage({ type: 'result', ...results });
+      }
     });
   });
 
@@ -140,8 +170,9 @@ async function runTestWithSeparateIndex(
   poolUtil: OpfsSAHPoolUtil,
   source: string,
   batchSize: number,
-  useTriggers?: boolean
-): Promise<{ insertDur: number; queryDur: number }> {
+  useTriggers?: boolean,
+  saveFile?: boolean
+): Promise<{ insertDur: number; queryDur: number; db: DB }> {
   const db = new poolUtil.OpfsSAHPoolDb('/sqlite-test');
   try {
     db.exec('PRAGMA locking_mode = exclusive');
@@ -211,12 +242,16 @@ async function runTestWithSeparateIndex(
     );
     const queryDur = performance.now() - queryStart;
 
-    // Tidy up
-    db.exec(['drop table readings;', 'drop table words']);
+    if (!saveFile) {
+      // Tidy up
+      db.exec(['drop table readings;', 'drop table words']);
+    }
 
-    return { insertDur, queryDur };
+    return { insertDur, queryDur, db };
   } finally {
-    db.close();
+    if (!saveFile) {
+      db.close();
+    }
     await poolUtil.wipeFiles();
   }
 }
